Moves short_url service to `services.short_url.py`
