+++ main.py
@@ -1,5 +1,3 @@
-import secrets
-import string
 from contextlib import asynccontextmanager
 from typing import AsyncGenerator
 
@@ -12,6 +10,7 @@ from config import settings
 from database import create_tables, get_db
 from models import Link
 from schemas import LongUrlAccept, LongUrlReturn, ShortUrlReturn
+from services.short_url import create_short_url
 
 app: FastAPI = FastAPI(title=settings.app_name)
 
@@ -20,33 +19,6 @@ app: FastAPI = FastAPI(title=settings.app_name)
 async def lifespan(app: FastAPI) -> AsyncGenerator:
     await create_tables()
     yield
-
-
-""" Short url service """
-
-
-def generate_slug(length: int) -> str:
-    base62: str = string.ascii_letters + string.digits
-    return "".join(secrets.choice(seq=base62) for _ in range(length))
-
-
-async def generate_unique_slug(db: AsyncSession) -> str:
-    while True:
-        slug: str = generate_slug(settings.slug_length)
-        slug_exists: Link | None = await db.scalar(
-            select(Link).where(Link.slug == slug)
-        )
-        if not slug_exists:
-            return slug
-
-
-async def create_short_url(db: AsyncSession, long_url: str) -> str:
-    slug: str = await generate_unique_slug(db)
-    link = Link(slug=slug, long_url=long_url)
-    db.add(link)
-    await db.commit()
-    await db.refresh(link)
-    return f"{settings.base_url}{slug}"
 
 
 class NoMatchingSlugError(Exception):
